[refactor] 일기 생성 API 코드 refectoring
