+++ src/main/java/Opensource_SW_Project/Project/converter/TalkConverter.java
@@ -1,9 +1,7 @@
 package Opensource_SW_Project.Project.converter;
 
 import Opensource_SW_Project.Project.domain.Talk;
-import Opensource_SW_Project.Project.domain.User;
-import Opensource_SW_Project.Project.web.dto.TalkResponseDTO;
-import Opensource_SW_Project.Project.web.dto.UserResponseDTO;
+import Opensource_SW_Project.Project.web.dto.Talk.TalkResponseDTO;
 
 public class TalkConverter {
     public static TalkResponseDTO.CreateTalkResultDTO toCreateTalkResultDTO(Talk talk) {
