+++ src/main/java/Opensource_SW_Project/Project/converter/UserConverter.java
@@ -1,7 +1,7 @@
 package Opensource_SW_Project.Project.converter;
 
 import Opensource_SW_Project.Project.domain.User;
-import Opensource_SW_Project.Project.web.dto.UserResponseDTO;
+import Opensource_SW_Project.Project.web.dto.User.UserResponseDTO;
 
 public class UserConverter {
     public static UserResponseDTO.CreateUserResultDTO toCreateUserResultDTO(User user) {
