+++ src/main/java/Opensource_SW_Project/Project/service/ChatgptApiService/ChatgptApiCommandService.java
@@ -2,7 +2,7 @@ package Opensource_SW_Project.Project.service.ChatgptApiService;
 
 import Opensource_SW_Project.Project.web.dto.Talk.TalkRequestDTO;
 
-public interface ChatgptApiService { // ChatGPT API부르는 함수 만들어 공통으로 쓰기 <- 2개의 api 응답을 합치는데 쓰일 것임
+public interface ChatgptApiCommandService { // ChatGPT API부르는 함수 만들어 공통으로 쓰기 <- 2개의 api 응답을 합치는데 쓰일 것임
 
     // userPrompt 받아오기
     String getUserPrompt(TalkRequestDTO.CreateMessageRequestDTO request);
