+++ src/main/java/Opensource_SW_Project/Project/service/DiaryService/DiaryCommandService.java
@@ -1,8 +1,8 @@
 package Opensource_SW_Project.Project.service.DiaryService;
 
 import Opensource_SW_Project.Project.domain.Diary;
-import Opensource_SW_Project.Project.web.dto.DiaryRequestDTO;
-import Opensource_SW_Project.Project.web.dto.TalkRequestDTO;
+import Opensource_SW_Project.Project.web.dto.Diary.DiaryRequestDTO;
+import Opensource_SW_Project.Project.web.dto.Talk.TalkRequestDTO;
 
 public interface DiaryCommandService {
     String createDiarySystemPrompt(Long userId);
