+++ src/main/java/Opensource_SW_Project/Project/web/controller/DiaryController.java
@@ -4,9 +4,13 @@ import Opensource_SW_Project.Project.apiPayload.ApiResponse;
 import Opensource_SW_Project.Project.apiPayload.code.status.SuccessStatus;
 import Opensource_SW_Project.Project.converter.DiaryConverter;
 import Opensource_SW_Project.Project.domain.Diary;
-import Opensource_SW_Project.Project.service.ChatgptApiService.ChatgptApiService;
+import Opensource_SW_Project.Project.service.ChatgptApiService.ChatgptApiCommandService;
 import Opensource_SW_Project.Project.service.DiaryService.DiaryCommandService;
-import Opensource_SW_Project.Project.web.dto.*;
+import Opensource_SW_Project.Project.web.dto.ChatGPT.ChatGPTRequestDTO;
+import Opensource_SW_Project.Project.web.dto.ChatGPT.ChatGPTResponseDTO;
+import Opensource_SW_Project.Project.web.dto.Diary.DiaryRequestDTO;
+import Opensource_SW_Project.Project.web.dto.Diary.DiaryResponseDTO;
+import Opensource_SW_Project.Project.web.dto.Talk.TalkRequestDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
@@ -24,7 +28,7 @@ import org.springframework.web.client.RestTemplate;
 @RequestMapping("/diary")
 @Slf4j
 public class DiaryController {
-    private final ChatgptApiService chatgptApiService;
+    private final ChatgptApiCommandService chatgptApiService;
     private final DiaryCommandService diaryCommandService;
 
     @Value("${openai.model}")
@@ -38,14 +42,18 @@ public class DiaryController {
 
     // 일기 생성
     @PostMapping()
+    @Operation(
+            summary = "일기 생성 API"
+            , description = "일기를 생성합니다. Param으로 userId과, RequestBody에 대화내역 talkId와 userPrompt를 입력하세요"
+            , security = @SecurityRequirement(name = "accessToken")
+    )
     public ApiResponse<DiaryResponseDTO.CreateDiaryResultDTO> createDiary(
             @RequestParam(name = "userId")Long userId,
             @RequestBody TalkRequestDTO.CreateMessageRequestDTO request
             ){
-        String userPrompt1 = diaryCommandService.createDiarySystemPrompt(userId);
-        String userPrompt2 = chatgptApiService.getHistorytalk(userId, request);
-
-        String userPrompt = userPrompt1 + userPrompt2;
+        String DiarySystemPrompt = diaryCommandService.createDiarySystemPrompt(userId);
+        String userHistoryTalk = chatgptApiService.getHistorytalk(userId, request);
+        String userPrompt = DiarySystemPrompt + userHistoryTalk;
 
         String systemPrompt = "";
         // request를 api로 보내 chatGPT응답받기
@@ -56,7 +64,7 @@ public class DiaryController {
         Diary newDiary = diaryCommandService.saveDiary(userId, request, chatGPTResponse.getChoices().get(0).getMessage().getContent());
 
         return ApiResponse.onSuccess(
-                SuccessStatus.MESSAGE_OK,
+                SuccessStatus.DIARY_OK,
                 DiaryConverter.toCreateDiaryResultDTO(
                         newDiary
                 )
@@ -68,7 +76,7 @@ public class DiaryController {
     @PatchMapping("/{diaryId}")
     @Operation(
             summary = "일기 수정 API"
-            , description = "일기를 수정합니다. Path variable로 diaryId를 입력 받고, RequestBody에 수정할 일기 content를 입력하세요"
+            , description = "일기를 수정합니다. Path variable로 diaryId를 입력 받고, RequestBody에 작성자 userId와 수정할 일기 content를 입력하세요"
             , security = @SecurityRequirement(name = "accessToken")
     )
     public ApiResponse<DiaryResponseDTO.UpdateDiaryResultDTO> updateDiary(
