+++ src/main/java/Opensource_SW_Project/Project/web/controller/TalkController.java
@@ -3,13 +3,9 @@ package Opensource_SW_Project.Project.web.controller;
 import Opensource_SW_Project.Project.apiPayload.ApiResponse;
 import Opensource_SW_Project.Project.apiPayload.code.status.SuccessStatus;
 import Opensource_SW_Project.Project.converter.TalkConverter;
-import Opensource_SW_Project.Project.converter.UserConverter;
 import Opensource_SW_Project.Project.domain.Talk;
-import Opensource_SW_Project.Project.domain.User;
 import Opensource_SW_Project.Project.service.TalkService.TalkCommandService;
-import Opensource_SW_Project.Project.service.UserService.UserCommandService;
-import Opensource_SW_Project.Project.web.dto.TalkResponseDTO;
-import Opensource_SW_Project.Project.web.dto.UserResponseDTO;
+import Opensource_SW_Project.Project.web.dto.Talk.TalkResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
