+++ src/main/java/Opensource_SW_Project/Project/web/dto/ChatGPT/ChatGPTRequestDTO.java
@@ -1,5 +1,6 @@
-package Opensource_SW_Project.Project.web.dto;
+package Opensource_SW_Project.Project.web.dto.ChatGPT;
 
+import Opensource_SW_Project.Project.web.dto.Message.MessageDTO;
 import lombok.Data;
 
 import java.util.ArrayList;
