+++ src/main/java/Opensource_SW_Project/Project/web/dto/ChatgptApi/ChatgptApiRequestDTO.java
@@ -1,4 +1,4 @@
-package Opensource_SW_Project.Project.web.dto;
+package Opensource_SW_Project.Project.web.dto.ChatgptApi;
 
 import lombok.*;
