+++ src/main/java/Opensource_SW_Project/Project/web/dto/Message/MessageDTO.java
@@ -1,4 +1,4 @@
-package Opensource_SW_Project.Project.web.dto;
+package Opensource_SW_Project.Project.web.dto.Message;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
