+++ src/main/java/Opensource_SW_Project/Project/web/dto/User/UserRequestDTO.java
@@ -1,4 +1,4 @@
-package Opensource_SW_Project.Project.web.dto;
+package Opensource_SW_Project.Project.web.dto.User;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
